requirements.txt
# Trading APIs
alpaca-trade-api==3.2.0
alpaca-py==0.30.1
finnhub-python==2.4.19
PyYAML>=6.0.1

# Data processing
pandas==2.2.2
numpy==1.26.4

# Configuration
python-dotenv==1.0.1

# Notifications
requests==2.32.3

# Database
# sqlite3 is built into Python

# Development and testing
pytest==8.2.2
pytest-cov==5.0.0

